+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-12-22
+
+- `SculptParams::from_mesh_graph()` now takes a parameter `min_edge_length`.
+- Updated dependencies `mesh-graph` and `rerun`.
+
 ## [0.4.1] - 2025-10-08
 
 - Fixed BVH